updated name space, naming of feature file and steps
