+++ platform/tests/EducationBenchmarking.Platform.ApiTests/EducationBenchmarking.Platform.ApiTests.csproj
@@ -6,7 +6,6 @@
         <ImplicitUsings>enable</ImplicitUsings>
         <IsPackable>false</IsPackable>
         <IsTestProject>true</IsTestProject>
-        <RootNamespace>EducationBenchmarking.platform.Api.Tests</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
