+++ platform/tests/EducationBenchmarking.Platform.ApiTests/Hooks/Hook.cs
@@ -1,0 +1,7 @@
+namespace EducationBenchmarking.Platform.ApiTests.Hooks
+{
+    [Binding]
+    public class Hooks
+    {
+    }
+}
