+++ platform/tests/EducationBenchmarking.Platform.ApiTests/appsettings.json
