--- platform/tests/EducationBenchmarking.platform.ApiTests/Hooks/Hook.cs
@@ -1,7 +1,0 @@
-namespace EducationBenchmarking.platfrom.Api.Tests.Hooks
-{
-    [Binding]
-    public class Hooks
-    {
-    }
-}
